+++ t/001-management.t
@@ -144,9 +144,9 @@ Could not resolve GET /foobar - nil
 exposes boot function
 --- main_config
 env THREESCALE_PORTAL_ENDPOINT=http://localhost:$TEST_NGINX_SERVER_PORT/config/;
+env RESOLVER=127.0.0.1:1953;
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
-  resolver 127.0.0.1:1953 ipv6=off;
   init_by_lua_block {
       require('configuration_loader').save({ services = { { id = 42 } } })
   }
@@ -159,7 +159,7 @@ POST /boot
 --- error_code: 200
 --- udp_listen: 1953
 --- udp_reply eval
-$::dns->("localhost", "127.0.0.1")
+$::dns->("localhost", "127.0.0.1", 60)
 --- no_error_log
 [error]
 
@@ -167,9 +167,9 @@ $::dns->("localhost", "127.0.0.1")
 keeps the same configuration
 --- main_config
 env THREESCALE_PORTAL_ENDPOINT=http://localhost:$TEST_NGINX_SERVER_PORT/config/;
+env RESOLVER=127.0.0.1:1953;
 --- http_config
   lua_package_path "$TEST_NGINX_LUA_PATH";
-  resolver 127.0.0.1:1953 ipv6=off;
   init_by_lua_block {
       require('configuration_loader').save({ services = { { id = 42 } } })
   }
@@ -187,7 +187,7 @@ POST /test
 --- error_code: 200
 --- udp_listen: 1953
 --- udp_reply eval
-$::dns->("localhost", "127.0.0.1")
+$::dns->("localhost", "127.0.0.1", 60)
 --- no_error_log
 [error]