+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use stale DNS cache when there is a query in progress for that record [PR #260](https://github.com/3scale/apicast/pull/260)
 - Bump s2i-openresty to 1.11.2.2-2 [PR #260](https://github.com/3scale/apicast/pull/260)
 - Echo API on port 8081 listens accepts any Host [PR #268](https://github.com/3scale/apicast/pull/268)
+- Always use DNS search scopes [PR #271](https://github.com/3scale/apicast/pull/271)
 
 ### Added